Reorganise legacy tests. Add new state dump.
