+++ tests/legacy/README.md
@@ -1,0 +1,5 @@
+Tests dealing with legacy data.
+
+Legacy data may not have all the same data points as the current code is generating.
+
+Test out that functionality, like statistics calculations do not crash, even if some data is missing.
