+++ tradeexecutor/state/trade.py
@@ -826,7 +826,7 @@ class TradeExecution:
         executed_quantity = self.executed_quantity or 0
         executed_price = self.executed_price or 0
 
-        return abs(float(executed_quantity) * (executed_price))
+        return abs(float(executed_quantity) * float(executed_price))
 
     def get_planned_value(self) -> USDollarAmount:
         """How much we plan to swap in this trade."""
